separate component to another file
