+++ frontend/components/register/SelectRoleForm.js
@@ -1,22 +1,8 @@
 import React from 'react';
 import { Button, Form, Row } from 'antd';
 
-import SelectionCard from '../SelectionCard';
+import SelectRoleInput from '../SelectRoleInput';
 import { USER_TYPE } from '../../config/UserType.config';
-
-const SelectRoleInput = ({ value , onChange }) => {
-  const onClick = role => {
-    if (onChange) {
-      onChange(role);
-    }
-  }
-  return (
-    <Row className="flex justify-around w-full">
-      <SelectionCard checked={value==USER_TYPE.TRAINER} imageUrl="/trainer.svg" description="Trainer" onClick={() => onClick(USER_TYPE.TRAINER)}/>
-      <SelectionCard checked={value==USER_TYPE.TRAINEE} imageUrl="/trainee.svg" description="Trainee" onClick={() => onClick(USER_TYPE.TRAINEE)}/>
-    </Row>
-  );
-};
 
 // NOTE: draft version
 const SelectRoleForm = ({ getState, setState, size, current, prev, next }) => {
